requirements.txt
streamlit>=1.38
streamlit-image-coordinates>=0.2.0
torch>=2.0
transformers>=4.40
pillow>=10.0
numpy>=1.24
matplotlib>=3.7
imageio>=2.31
imageio-ffmpeg>=0.4.9
opencv-python>=4.8
